S7/feature/finance

+++ database/seeds/DoctorsTableSeeder.php
@@ -5,7 +5,10 @@
  * @author Alexander Zagovorichev <zagovorichev@gmail.com>
  */
 
+use App\City;
 use App\Doctor;
+use App\Role;
+use App\User;
 use Illuminate\Database\Seeder;
 
 class DoctorsTableSeeder extends Seeder
@@ -17,10 +20,33 @@ class DoctorsTableSeeder extends Seeder
      */
     public function run()
     {
-        Doctor::truncate();
-        factory(Doctor::class, 10)->create();
-        factory(Doctor::class)->create([
-            'user_id' => \App\User::where('email', 'doctor@mail.com')->first()->id
-        ]);
+        if (App::environment('production') && Doctor::all()->count()) {
+            return;
+        } elseif (App::environment('production')) {
+            Doctor::truncate();
+
+            factory(Doctor::class, 10)->create([
+                'city_id' => function() {
+                    return factory(City::class)->create()->id;
+                }
+            ]);
+
+            $loginRoleId = Role::firstOrCreate(['title' => Role::ROLE_LOGIN])->id;
+            $doctorRoleId = Role::firstOrCreate(['title' => Role::ROLE_DOCTOR])->id;
+
+            $doctor = factory(\App\Doctor::class)->create([
+                'name' => 'Doctor Aibolit',
+                'user_id' => function() {
+                    return factory(User::class)->create([
+                        'email' => 'doctor@mail.com',
+                        'name' => 'Peter',
+                    ]);
+                },
+                'city_id' => function() {
+                    return factory(City::class)->create()->id;
+                }
+            ]);
+            $doctor->user->roles()->attach([$loginRoleId, $doctorRoleId]);
+        }
     }
 }
